+++ lib/src/widgets/animated_focus_light.dart
@@ -29,6 +29,7 @@ class AnimatedFocusLight extends StatefulWidget {
   final bool pulseEnable;
   final bool rootOverlay;
   final ImageFilter? imageFilter;
+  final int initialFocus;
 
   const AnimatedFocusLight({
     Key? key,
@@ -49,6 +50,7 @@ class AnimatedFocusLight extends StatefulWidget {
     this.imageFilter,
     this.pulseEnable = true,
     this.rootOverlay = false,
+    this.initialFocus = 0,
   })  : assert(targets.length > 0),
         super(key: key);
 
@@ -73,11 +75,12 @@ abstract class AnimatedFocusLightState extends State<AnimatedFocusLight>
   double _sizeCircle = 100;
   int _currentFocus = 0;
   double _progressAnimated = 0;
-  bool _goNext = true;
+  int nextIndex = 0;
 
   @override
   void initState() {
     super.initState();
+    _currentFocus = widget.initialFocus;
     _targetFocus = widget.targets[_currentFocus];
     _controller = AnimationController(
       vsync: this,
@@ -102,20 +105,31 @@ abstract class AnimatedFocusLightState extends State<AnimatedFocusLight>
 
   void next() => _tapHandler();
 
-  void previous() => _tapHandler(goNext: false);
+  void previous() {
+    nextIndex--;
+    _revertAnimation();
+  }
+
+  void goTo(int index) {
+    nextIndex = index;
+    _revertAnimation();
+  }
 
   Future _tapHandler({
-    bool goNext = true,
     bool targetTap = false,
     bool overlayTap = false,
   }) async {
+    nextIndex++;
     if (targetTap) {
       await widget.clickTarget?.call(_targetFocus);
     }
     if (overlayTap) {
       await widget.clickOverlay?.call(_targetFocus);
     }
+    return _revertAnimation();
   }
+
+  Future _revertAnimation();
 
   Future _tapHandlerForPosition(TapDownDetails tapDetails) async {
     await widget.clickTargetWithTapPosition?.call(_targetFocus, tapDetails);
@@ -168,23 +182,13 @@ abstract class AnimatedFocusLightState extends State<AnimatedFocusLight>
     _controller.forward();
   }
 
-  void _nextFocus() {
-    if (_currentFocus >= widget.targets.length - 1) {
+  void _goToFocus(int index) {
+    if (index >= 0 && index < widget.targets.length) {
+      _currentFocus = index;
+      _runFocus();
+    } else {
       _finish();
-      return;
     }
-    _currentFocus++;
-
-    _runFocus();
-  }
-
-  void _previousFocus() {
-    if (_currentFocus <= 0) {
-      _finish();
-      return;
-    }
-    _currentFocus--;
-    _runFocus();
   }
 
   void _finish() {
@@ -287,18 +291,8 @@ class AnimatedStaticFocusLightState extends AnimatedFocusLightState {
   }
 
   @override
-  Future _tapHandler({
-    bool goNext = true,
-    bool targetTap = false,
-    bool overlayTap = false,
-  }) async {
-    await super._tapHandler(
-      goNext: goNext,
-      targetTap: targetTap,
-      overlayTap: overlayTap,
-    );
-    safeSetState(() => _goNext = goNext);
-    _controller.reverse();
+  Future _revertAnimation() {
+    return _controller.reverse();
   }
 
   @override
@@ -307,15 +301,11 @@ class AnimatedStaticFocusLightState extends AnimatedFocusLightState {
       widget.focus?.call(_targetFocus);
     }
     if (status == AnimationStatus.dismissed) {
-      if (_goNext) {
-        _nextFocus();
-      } else {
-        _previousFocus();
-      }
+      _goToFocus(nextIndex);
     }
 
     if (status == AnimationStatus.reverse) {
-      widget.removeFocus!();
+      widget.removeFocus?.call();
     }
   }
 }
@@ -412,22 +402,12 @@ class AnimatedPulseFocusLightState extends AnimatedFocusLightState {
   }
 
   @override
-  Future _tapHandler({
-    bool goNext = true,
-    bool targetTap = false,
-    bool overlayTap = false,
-  }) async {
-    await super._tapHandler(
-      goNext: goNext,
-      targetTap: targetTap,
-      overlayTap: overlayTap,
-    );
+  Future _revertAnimation() {
     safeSetState(() {
-      _goNext = goNext;
       _initReverse = true;
     });
 
-    _controllerPulse.reverse(from: _controllerPulse.value);
+    return _controllerPulse.reverse(from: _controllerPulse.value);
   }
 
   @override
@@ -450,11 +430,7 @@ class AnimatedPulseFocusLightState extends AnimatedFocusLightState {
         _finishFocus = false;
         _initReverse = false;
       });
-      if (_goNext) {
-        _nextFocus();
-      } else {
-        _previousFocus();
-      }
+      _goToFocus(nextIndex);
     }
 
     if (status == AnimationStatus.reverse) {