+++ lib/src/widgets/tutorial_coach_mark_widget.dart
@@ -34,6 +34,7 @@ class TutorialCoachMarkWidget extends StatefulWidget {
     this.rootOverlay = false,
     this.showSkipInLastTarget = false,
     this.imageFilter,
+    this.initialFocus = 0,
   })  : assert(targets.length > 0),
         super(key: key);
 
@@ -61,6 +62,7 @@ class TutorialCoachMarkWidget extends StatefulWidget {
   final bool rootOverlay;
   final bool showSkipInLastTarget;
   final ImageFilter? imageFilter;
+  final int initialFocus;
 
   @override
   TutorialCoachMarkWidgetState createState() => TutorialCoachMarkWidgetState();
@@ -80,6 +82,7 @@ class TutorialCoachMarkWidgetState extends State<TutorialCoachMarkWidget>
         children: <Widget>[
           AnimatedFocusLight(
             key: _focusLightKey,
+            initialFocus: widget.initialFocus,
             targets: widget.targets,
             finish: widget.finish,
             paddingFocus: widget.paddingFocus,
@@ -253,31 +256,31 @@ class TutorialCoachMarkWidgetState extends State<TutorialCoachMarkWidget>
       return const SizedBox.shrink();
     }
 
-  Widget animatedWidget = AnimatedOpacity(
-    opacity: showContent ? 1 : 0,
-    duration: const Duration(milliseconds: 300),
-    child: InkWell(
-      onTap: skip,
-      child: Padding(
-        padding: const EdgeInsets.all(20.0),
-        child: IgnorePointer(
-          ignoringSemantics: false,
-          child: widget.skipWidget ??
-              Text(
-                widget.textSkip,
-                style: widget.textStyleSkip,
-              ),
+    Widget animatedWidget = AnimatedOpacity(
+      opacity: showContent ? 1 : 0,
+      duration: const Duration(milliseconds: 300),
+      child: InkWell(
+        onTap: skip,
+        child: Padding(
+          padding: const EdgeInsets.all(20.0),
+          child: IgnorePointer(
+            child: widget.skipWidget ??
+                Text(
+                  widget.textSkip,
+                  style: widget.textStyleSkip,
+                ),
+          ),
         ),
       ),
-    ),
-  );
+    );
 
-  return Align(
-    alignment: currentTarget?.alignSkip ?? widget.alignSkip,
-    child: (widget.useSafeArea)
-        ? SafeArea(child: animatedWidget)
-        : animatedWidget,
-  );
+    return Align(
+      alignment: currentTarget?.alignSkip ?? widget.alignSkip,
+      child: (widget.useSafeArea)
+          ? SafeArea(child: animatedWidget)
+          : animatedWidget,
+    );
+  }
 
   @override
   void skip() => widget.onClickSkip?.call();
@@ -287,4 +290,6 @@ class TutorialCoachMarkWidgetState extends State<TutorialCoachMarkWidget>
 
   @override
   void previous() => _focusLightKey.currentState?.previous();
+
+  void goTo(int index) => _focusLightKey.currentState?.goTo(index);
 }